[MAYID P5 16/09] S3Relay: dem theo ma may qua MOI ket noi GameServer + kep 255
CountLoginByHWID cu duyet m_mapIp2Connect (khoa theo DWORD IP): GameServer thu hai cung IP ghi de
muc cua GameServer thu nhat, OnClientConnectClose erase(ip) khi mot trong hai dong => co luc khong
GameServer nao cua IP do duoc dem. Nay duyet m_mapId2Connect cua CNetServer qua ham moi
CNetServer::ForEachConnect (giong CNetServer::Route), dem bang int va kep o 255 vi num_login tren
day la BYTE (256 phien cung ma quay ve 0 = qua nguong), co rTRACE khi kep.
CHostConnect::CountLoginByHWID doi BYTE -> int. Dich Release|Win32: 0 loi C.

+++ Sources/MultiServer/S3Relay/HostConnect.cpp
@@ -718,11 +718,11 @@ BOOL CHostConnect::FindPlayerByRole(const std::_tstring& role, std::_tstring* pA
 	return TRUE;
 }
 
-BYTE CHostConnect::CountLoginByHWID(const std::_tstring& hwid)
+int CHostConnect::CountLoginByHWID(const std::_tstring& hwid)
 {
 	DUMMY_AUTOLOCKREAD(m_lockPlayer);
 	
-	BYTE mCount = 0;
+	int mCount = 0;
 	for (PARAMMAP::iterator it = m_mapParam.begin(); it != m_mapParam.end(); it++)
 	{
 		const PARAMINFO& infoParam = (*it).second;
